Move common test fixtures to dstack package
